Changes to MainScreen and Saving via AppEnvironment
For MainScreen and it's Logic class
MainScreen now extends Screen. When the Screen is closes, the application is automatically saved.

Moved methods around and made code more cleaner and better style.

Method to handle closing a screen

For AppEnvironment
Added a method to save the application.

Now has a Setup attribute, which is nice bc it remembers where the app is saved (working directory) making things super easy to save.